WATER-AUDIT (lane B): the beach jitter, the nature compare and the ocean constant are the reference's floats
The audit of WATER1's ocean-band fix found the claim true and the fix
right, with three residues in the same arithmetic class, taken here:

- UMRandom.nextFloatRange was `nextFloat() * (max - min) + min` in
  double; Unity.Mathematics.Random.NextFloat(min, max) rounds every
  operation to float32. Measured: 17% of draws differed by a float32
  ulp, and through the beach jitter's fround(40 + j) the tile threshold
  itself moved for 1% of corners. Float32 per operation now.
- terrainNature.js compared the same DFU expression in double where the
  tile job now compares in float32; the twin compares agree.
- SCALED_OCEAN_ELEVATION is Math.fround(3.4 * 8), the float 3.4f * 8
  is in C#, so the port carries one value of the ocean elevation and
  the tile job reads it directly (the _F32 twins are gone). The stored
  normalised sea is the same float either way, pinned.

+++ src/world/terrainNature.js
@@ -133,7 +133,8 @@ export function layoutNature(heightmapData, tilemapData, opts) {
       const hx = Math.min(hDim - 1, Math.max(0, Math.trunc(hDim * (x / tDim))));
       const hy = Math.min(hDim - 1, Math.max(0, Math.trunc(hDim * (y / tDim))));
       // x & y swapped in heightmap, verbatim; unscaled height vs beach.
-      const height = heightmapData[hy + hx * hDim] * MAX_TERRAIN_HEIGHT;
+      // WATER-AUDIT: float32, as the tile job's twin of this line is.
+      const height = Math.fround(heightmapData[hy + hx * hDim] * MAX_TERRAIN_HEIGHT);
       if (height < beachLine) continue;
 
       const record = rng.nextIntRange(1, 32);
